Move most dictionary manipulation to fsdp_optim_utils.py
